22. Design change (51)

+++ common/models/CategoryCityContent.php
@@ -36,6 +36,7 @@ class CategoryCityContent extends ActiveRecord {
 	public $cities = [];
 	public $cities_options = [];
 	public $preview;
+	private $image_updated = false;
 	
 	/**
 	 * @inheritdoc
@@ -198,9 +199,8 @@ class CategoryCityContent extends ActiveRecord {
 	
 	public function beforeSave($insert){
 		
-		if(!$insert){
-			$id = intval(Yii::$app->request->get('id'));
-			$this->saveImages($id, $insert);
+		if(!$this->image_updated && !$insert){
+			$this->saveImages($insert);
 		}
 		
 		return parent::beforeSave($insert);
@@ -210,15 +210,14 @@ class CategoryCityContent extends ActiveRecord {
 		parent::afterSave($insert, $changedAttributes);
 
 		if($insert){
-			
 			$this->saveImages($this->category_id, $insert);
 		}
 	}
 	
-	private function saveImages($id = 0, $insert){
-		if($id == 0) return;
+	private function saveImages($insert){
+		#Yii::info(__CLASS__.' > '.__FUNCTION__, 'past55');
 		
-		$dir = Yii::getAlias('@3c_images').'/'.$id;
+		$dir = Yii::getAlias('@3c_images').'/'.$this->id;
 		
 		if(!is_dir($dir)){
 			FileHelper::createDirectory($dir, 0777);
@@ -226,24 +225,42 @@ class CategoryCityContent extends ActiveRecord {
 		}
 		
 		if($file = UploadedFile::getInstance($this, 'preview')){
-			#VarDumper::dump($file, 10, 1); exit;
-			if(file_exists($dir.'/'.$this->image) && is_file($dir.'/'.$this->image)){
+			#Yii::info($file, 'past55');
+			
+			if(!$insert && !empty($this->image) && file_exists($dir.'/'.$this->image) && is_file($dir.'/'.$this->image)){
 				FileHelper::unlink($dir.'/'.$this->image);
-			}
-			if(file_exists($dir.'/thumbs/'.$this->image) && is_file($dir.'/thumbs/'.$this->image)){
-				FileHelper::unlink($dir.'/thumbs/'.$this->image);
+				
+				foreach(Yii::$app->params['image_sizes'] as $name => $size){
+					$replace_fragment = [];
+					if(!is_null($size['w'])) $replace_fragment[] = $size['w'];
+					if(!is_null($size['h'])) $replace_fragment[] = $size['h'];
+					$replace_fragment = implode('_', $replace_fragment);
+					
+					$image = str_replace('.'.$file->extension, '_'.$replace_fragment.'.'.$file->extension, $this->image);
+					
+					if(file_exists($dir.'/thumbs/'.$image) && is_file($dir.'/thumbs/'.$image)){
+						FileHelper::unlink($dir.'/thumbs/'.$image);
+					}
+				}
 			}
 			
-			$this->image = $id.'_'.time().'_'.rand(137, 999).'.'.$file->extension;
+			$this->image = $this->id.'_'.time().'_'.rand(137, 999).'.'.$file->extension;
 			
 			$file->saveAs($dir.'/'.$this->image);
-
+			
 			foreach(Yii::$app->params['image_sizes'] as $name => $size){
+				$replace_fragment = [];
+				if(!is_null($size['w'])) $replace_fragment[] = $size['w'];
+				if(!is_null($size['h'])) $replace_fragment[] = $size['h'];
+				$replace_fragment = implode('_', $replace_fragment);
+				
 				$image = Yii::$app->image->load($dir.'/'.$this->image);
 				$image->background('#fff', 0);
-				$image->resize($size, null, Image::INVERSE);
-				$image->save($dir.'/thumbs/'.str_replace('.'.$file->extension, '_'.$size.'.'.$file->extension, $this->image), 90);
+				$image->resize($size['w'], $size['h'], $size['crop']);
+				$image->save($dir.'/thumbs/'.str_replace('.'.$file->extension, '_'.$replace_fragment.'.'.$file->extension, $this->image), 90);
 			}
+			
+			$this->image_updated = true;
 		}
 		
 		if($insert){
